Moved search button to the propper menu.
