+++ app/src/main/java/de/zalando/zmon/fragment/TeamListFragment.java
@@ -1,18 +1,23 @@
 package de.zalando.zmon.fragment;
 
-import android.app.Activity;
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.AdapterView;
-import android.widget.ListView;
+import java.util.List;
+
 import de.zalando.zmon.R;
 import de.zalando.zmon.adapter.TeamListAdapter;
 import de.zalando.zmon.persistence.Team;
 
-import java.util.List;
+import android.app.Activity;
+
+import android.os.Bundle;
+
+import android.support.v4.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import android.widget.AdapterView;
+import android.widget.ListView;
 
 public class TeamListFragment extends Fragment {
 
@@ -60,4 +65,5 @@ public class TeamListFragment extends Fragment {
 
         teamList.setAdapter(adapter);
     }
+
 }
