+++ app/src/main/res/menu/menu_remote_team_list.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+      xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item android:id="@+id/search_team"
+          android:enabled="true"
+          android:title="@string/app_name"
+          android:icon="@drawable/ic_search_white_24dp"
+          app:showAsAction="collapseActionView|ifRoom"
+          app:actionViewClass="android.support.v7.widget.SearchView" />
+
+</menu>
