+++ app/src/main/res/menu/menu_team_list.xml
@@ -2,13 +2,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <item android:id="@+id/search_team"
-          android:enabled="true"
-          android:title="@string/app_name"
-          android:icon="@drawable/ic_menu_search"
-          app:showAsAction="collapseActionView|ifRoom"
-          app:actionViewClass="android.support.v7.widget.SearchView" />
-
     <item
             android:id="@+id/add_team"
             android:enabled="true"
